KI-Projekt/requirements.txt
Flask
Werkzeug
transformers
easyocr
opencv-python
Pillow
numpy

